requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
reportlab
python-multipart
pytest
pytest-asyncio
httpx
